(Abstract) com_light, sys_light
