+++ Abstract/game.ts
@@ -13,6 +13,10 @@ export interface Game {
     Canvas: HTMLCanvasElement;
     Camera?: Camera;
 
+    // The rendering pipeline supports 8 lights.
+    LightPositions: Float32Array;
+    LightDetails: Float32Array;
+
     FrameReset(): void;
     FrameUpdate(delta: number): void;
 }
