+++ Abstract/world.ts
@@ -1,13 +1,16 @@
 import {Camera} from "./components/com_camera.js";
+import {Light} from "./components/com_light.js";
 import {Transform} from "./components/com_transform.js";
 
 const enum Component {
     Camera,
+    Light,
     Transform,
 }
 
 export const enum Has {
     Camera = 1 << Component.Camera,
+    Light = 1 << Component.Light,
     Transform = 1 << Component.Transform,
 }
 
@@ -17,5 +20,6 @@ export interface World {
 
     // Component data
     Camera: Array<Camera>;
+    Light: Array<Light>;
     Transform: Array<Transform>;
 }
